fc_phasing.py, move the code to src/py/phasing.py and leave src/py_scripts/fc_phasing.py as entry point fc_unzip.py, move previous shell script cmd line calls into pypeflow
